Fix penalty data type in Icpc scoreboard entry

+++ judgels-backends/uriel/uriel-api/src/main/java/judgels/uriel/api/contest/scoreboard/IcpcScoreboard.java
@@ -27,8 +27,8 @@ public interface IcpcScoreboard extends Scoreboard {
         int getRank();
         String getContestantJid();
         int getTotalAccepted();
-        int getTotalPenalties();
-        int getLastAcceptedPenalty();
+        long getTotalPenalties();
+        long getLastAcceptedPenalty();
         List<Integer> getAttemptsList();
         List<Long> getPenaltyList();
         List<IcpcScoreboardProblemState> getProblemStateList();
